requirements.txt
python-telegram-bot>=21.0
python-dotenv
requests
google-generativeai
twikit>=2.0
